split fragment.py from volume.py
